Update deps. and README, make download optional
Add new skippable tag to the provisioning playbook to avoid downloading
the Nomad tools binaries and look directly for the files under /tmp.
Also updated some download URLs and the path of the Podman socket in the
Nomad HCL template.
Finally enabled the option to clean all /tmp on the managed host on
boot.